offer-page-reviews component created
